Added permissions checking to conversationDialog actions (delete, like, reply?) Moved dahsboard and page statistics boxes to their own folder called widgets allowing us to re-use them on multiple pages Fixed issue with theme not applying on socialStackDialog Removed the loading of pages from the pageController since it is already loaded into the root portalHomeController Added the passing in socialPages from root portalHomeController to avoid re-looking up the social pages again in the socialStackDialogController Removed pageStatisticController code since its only looking up socialPages which are already in the root portalHomeController Fixed the placeholder text in the graphs to be sized according to the div's size so it should not be sized to fit within the div (before it exceeded it in the content analysis div/widget) Added pageSelector directive to ease the re-use of social page selectors with search capabilities... Need to add brand support soon.. Added socialPageStatistics panel in the  pageListcontroller Added new billing pricing page and styling that pops out.

+++ src/app/controllers/portal/Dashboard/conversationDialogController.js
@@ -23,10 +23,12 @@ define(['../../module'], controllers => {
                     open: {
                         page_id: socialPage
                     }
-                }
+                };
+
                 $scope.allowedActions = allowedActions;
                 console.log($scope.allowedActions);
                 $scope.postComment = reply => {
+                    if (!$scope.permissions.post_manage_reply) return Alert.error("Please upgrade to a plan that offers ability to reply.");
                     if (!reply.comment || !reply._id) return;
                     Action.postComment({reply_id: reply._id, reply: reply.comment},
                         data => {
@@ -73,6 +75,8 @@ define(['../../module'], controllers => {
                 };
 
                 $scope.deleteComment = reply => {
+                    if (!$scope.permissions.post_manage_delete) return Alert.error("Please upgrade to a plan that offers ability to delete.");
+
                     reply.deleted = true;
                     Action.deleteComment({reply_id: reply._id}, data => {
                         Alert.success("Deleted selected comment", 600);
